Restructuration de l'architecture des fichers et réglage d'un souci de redirection lorsque l'utilisateur se connecte
