+++ src/App.js
@@ -1,5 +1,4 @@
 import { BrowserRouter, Routes, Route, Navigate } from "react-router-dom";
-import Home from "./pages/home";
 import Test from "./pages/connect_page";
 import Statistiques from "./pages/statistique_page";
 import Fournisseur from "./pages/fournisseur_page";
@@ -7,7 +6,7 @@ import { AuthProvider, useAuthContext } from "./pages/AuthProvider";
 
 const ProtectedRoute = ({ children }) => {
     const { user } = useAuthContext();
-    return user ? children : <Navigate to="/home" />;
+    return user ? children : <Navigate to="/test" />;
 };
 
 function App() {
@@ -17,33 +16,27 @@ function App() {
                 <Routes>
                     {/* Route par défaut (accueil) */}
                     <Route path="/" element={<Navigate to="/test" />} />
-
-                    {/* Route publique */}
-                    <Route path="/home" element={<Home />} />
-
-                    {/* Routes protégées */}
                     <Route
                         path="/test"
                         element={
-
                                 <Test />
-
                         }
                     />
+                    {/* Routes accessible que après s'être connecté */}
                     <Route
                         path="/statistiques"
                         element={
-                           /* <ProtectedRoute>*/
+                            <ProtectedRoute>
                                 <Statistiques />
-                           /* </ProtectedRoute>*/
+                           </ProtectedRoute>
                         }
                     />
                     <Route
                         path="/fournisseur"
                         element={
-                            /*<ProtectedRoute>*/
+                            <ProtectedRoute>
                                 <Fournisseur />
-                           /* </ProtectedRoute>*/
+                           </ProtectedRoute>
                         }
                     />
                 </Routes>
