+++ src/components/FournisseurTable.js
@@ -1,0 +1,20 @@
+function FournisseurTable() {
+    return (
+        <div className="liste-fournisseur">
+            <table className="table_fournisseur">
+                <thead>
+                <tr>
+                    <th scope="col">Lorem ipsum</th>
+                    <th scope="col">Lorem ipsum</th>
+                    <th scope="col">Lorem ipsum</th>
+                </tr>
+                </thead>
+                <tbody>
+                {/* à compléter en backend */}
+                </tbody>
+            </table>
+        </div>
+    );
+}
+
+export default FournisseurTable;
